Basic Demo
All changes are in scene Level_0. The controls are wasd to move, right click to change camera angle, and click to control buttons (possible change for now). Press buttons A-G to strike a chord, toggle the b and # to make the note flat or sharp, and toggle the m to make the chord minor. The sample sounds do not sound good enough in my opinion, so I will try to find better ones online later. Have a try, and happy thanksgiving!